Combine fixer & analyzer tests and delete duplicate tests
